+++ crates/rig-core/src/test_utils/mod.rs
@@ -1,7 +1,13 @@
 //! Test utilities for deterministic completion-model tests.
 
 mod completion;
+mod embeddings;
+mod model_listing;
+mod pipeline;
 mod streaming;
 
 pub use completion::{MockCompletionModel, MockError, MockTurn};
+pub use embeddings::MockEmbeddingModel;
+pub use model_listing::MockModelLister;
+pub use pipeline::{Foo, MockPromptModel, MockVectorStoreIndex};
 pub use streaming::{MockResponse, MockStreamEvent};
